# Add common scalar reads to OrcInputStream