--- a/presto-orc/src/main/java/com/facebook/presto/orc/stream/OrcInputStream.java
+++ b/presto-orc/src/main/java/com/facebook/presto/orc/stream/OrcInputStream.java
@@ -18,6 +18,7 @@ import com.facebook.presto.memory.context.LocalMemoryContext;
 import com.facebook.presto.orc.OrcCorruptionException;
 import com.facebook.presto.orc.OrcDataSourceId;
 import com.facebook.presto.orc.OrcDecompressor;
+import com.facebook.presto.orc.metadata.OrcType.OrcTypeKind;
 import io.airlift.slice.ByteArrays;
 import io.airlift.slice.FixedLengthSliceInput;
 import io.airlift.slice.Slice;
@@ -30,10 +31,14 @@ import java.util.Optional;
 import static com.facebook.presto.orc.checkpoint.InputStreamCheckpoint.createInputStreamCheckpoint;
 import static com.facebook.presto.orc.checkpoint.InputStreamCheckpoint.decodeCompressedBlockOffset;
 import static com.facebook.presto.orc.checkpoint.InputStreamCheckpoint.decodeDecompressedOffset;
+import static com.facebook.presto.orc.stream.LongDecode.zigzagDecode;
 import static com.google.common.base.MoreObjects.toStringHelper;
 import static com.google.common.base.Preconditions.checkArgument;
 import static io.airlift.slice.SizeOf.SIZE_OF_DOUBLE;
 import static io.airlift.slice.SizeOf.SIZE_OF_FLOAT;
+import static io.airlift.slice.SizeOf.SIZE_OF_INT;
+import static io.airlift.slice.SizeOf.SIZE_OF_LONG;
+import static io.airlift.slice.SizeOf.SIZE_OF_SHORT;
 import static io.airlift.slice.Slices.EMPTY_SLICE;
 import static java.lang.Math.toIntExact;
 import static java.util.Objects.requireNonNull;
@@ -42,6 +47,9 @@ import static sun.misc.Unsafe.ARRAY_BYTE_BASE_OFFSET;
 public final class OrcInputStream
         extends InputStream
 {
+    private static final long VARINT_MASK = 0x8080_8080_8080_8080L;
+    private static final int MAX_VARINT_LENGTH = 10;
+
     private final OrcDataSourceId orcDataSourceId;
     private final FixedLengthSliceInput compressedSliceInput;
     private final Optional<OrcDecompressor> decompressor;
@@ -244,6 +252,153 @@ public final class OrcInputStream
         result = Math.min(available(), n - 1);
         position += toIntExact(result);
         return 1 + result;
+    }
+
+    public long readDwrfLong(OrcTypeKind type)
+            throws IOException
+    {
+        switch (type) {
+            case SHORT:
+                return read() | (read() << 8);
+            case INT:
+                return read() | (read() << 8) | (read() << 16) | (read() << 24);
+            case LONG:
+                return ((long) read()) |
+                        (((long) read()) << 8) |
+                        (((long) read()) << 16) |
+                        (((long) read()) << 24) |
+                        (((long) read()) << 32) |
+                        (((long) read()) << 40) |
+                        (((long) read()) << 48) |
+                        (((long) read()) << 56);
+            default:
+                throw new IllegalStateException();
+        }
+    }
+
+    public void skipDwrfLong(OrcTypeKind type, long items)
+            throws IOException
+    {
+        if (items == 0) {
+            return;
+        }
+        long bytes = items;
+        switch (type) {
+            case SHORT:
+                bytes *= SIZE_OF_SHORT;
+                break;
+            case INT:
+                bytes *= SIZE_OF_INT;
+                break;
+            case LONG:
+                bytes *= SIZE_OF_LONG;
+                break;
+            default:
+                throw new IllegalStateException();
+        }
+        skip(bytes);
+    }
+
+    public long readVarint(boolean signed)
+            throws IOException
+    {
+        long result = 0;
+        int shift = 0;
+        int available = available();
+        if (available >= 2 * Long.BYTES) {
+            long word = ByteArrays.getLong(buffer, position);
+            int count = 1;
+            boolean atEnd = false;
+            result = word & 0x7f;
+            if ((word & 0x80) != 0) {
+                long control = word >>> 8;
+                long mask = 0x7f << 7;
+                while (true) {
+                    word = word >>> 1;
+                    result |= word & mask;
+                    count++;
+                    if ((control & 0x80) == 0) {
+                        atEnd = true;
+                        break;
+                    }
+                    if (mask == 0x7fL << (7 * 7)) {
+                        break;
+                    }
+                    mask = mask << 7;
+                    control = control >>> 8;
+                }
+                if (!atEnd) {
+                    word = ByteArrays.getLong(buffer, position + 8);
+                    result |= (word & 0x7f) << 56;
+                    if ((word & 0x80) == 0) {
+                        count++;
+                    }
+                    else {
+                        result |= 1L << 63;
+                        count += 2;
+                    }
+                }
+            }
+            position += count;
+        }
+        else {
+            do {
+                if (available == 0) {
+                    advance();
+                    available = available();
+                    if (available == 0) {
+                        throw new OrcCorruptionException(orcDataSourceId, "End of stream in RLE Integer");
+                    }
+                }
+                available--;
+                result |= (long) (buffer[position] & 0x7f) << shift;
+                shift += 7;
+            }
+            while ((buffer[position++] & 0x80) != 0);
+        }
+        if (signed) {
+            return zigzagDecode(result);
+        }
+        else {
+            return result;
+        }
+    }
+
+    public void skipVarints(long items)
+            throws IOException
+    {
+        if (items == 0) {
+            return;
+        }
+
+        while (items > 0) {
+            items -= skipVarintsInBuffer(items);
+        }
+    }
+
+    private long skipVarintsInBuffer(long items)
+            throws IOException
+    {
+        if (available() == 0) {
+            advance();
+            if (available() == 0) {
+                throw new OrcCorruptionException(orcDataSourceId, "Unexpected end of stream");
+            }
+        }
+        long skipped = 0;
+        // If items to skip is > SIZE_OF_LONG it is safe to skip entire longs
+        while (items - skipped > SIZE_OF_LONG && available() > MAX_VARINT_LENGTH) {
+            long value = ByteArrays.getLong(buffer, position);
+            position += SIZE_OF_LONG;
+            long mask = (value & VARINT_MASK) ^ VARINT_MASK;
+            skipped += Long.bitCount(mask);
+        }
+        while (skipped < items && available() > 0) {
+            if ((buffer[position++] & 0x80) == 0) {
+                skipped++;
+            }
+        }
+        return skipped;
     }
 
     public double readDouble()