@oerling 2019-09-23
Thank you for the review.

I updated the PR. Do you mean we should split the commits now?

 

 

Release note?

 

A: Any example of usage for this?

 

 

 

I don't understand why you call sliceInput.setPosition(sliceInputPosition); - Would you explain?

 

 

A: This was the original from Elon. I take it the intent was not to have a side effect on the input. Another behavior would be to leave the slice input at end.

 

Neither exactly corresponds to the original, which delays the side effect on the input to the time of reading. This too could be reproduced if we wanted to.

 

Considering the use case, OrcInputStream takes ownership of the input and it can do the side effect at any or no time. So, leaving the stream at end is just as good and less code. There are no cases where any other reader from the slice input could/should coexist with the OrcInputStream, so on balance of safety leaving the stream at end could be seen as best.

 

 

 

 

 

 

 

 

 

 

 

From: Maria Basmanova <notifications@github.com> 
Sent: Monday, September 23, 2019 9:09 AM
To: prestodb/presto <presto@noreply.github.com>
Cc: oerling <erling@xs4all.nl>; Mention <mention@noreply.github.com>
Subject: Re: [prestodb/presto] Add common scalar reads to OrcInputStream (#13434)

 

@mbasmanova requested changes on this pull request.

@oerling <https://github.com/oerling>  Fantastic! This is such a powerful optimization and it applies to both default and Aria scans. I ran benchmarks for both scans for floats and doubles and I'm seeing 2-3x improvement. Thank you!

(Looks like our benchmarks don't cover DWRF longs. We should add these.)

---- with optimization ----
 
# Run complete. Total time: 00:03:25
 
 
 
Benchmark                             (typeSignature)  (withNulls)  Mode  Cnt  Score   Error  Units
 
BenchmarkSelectiveStreamReaders.read             real         true  avgt   20  0.047 ± 0.001   s/op
 
BenchmarkSelectiveStreamReaders.read             real        false  avgt   20  0.041 ± 0.002   s/op
 
BenchmarkSelectiveStreamReaders.read           double         true  avgt   20  0.061 ± 0.003   s/op
 
BenchmarkSelectiveStreamReaders.read           double        false  avgt   20  0.064 ± 0.004   s/op
 
 
 
 
 
 
 
# Run complete. Total time: 00:05:10
 
 
 
Benchmark                               (typeSignature)  (withNulls)  Mode  Cnt  Score   Error  Units
 
BenchmarkBatchStreamReaders.readBlocks             real      PARTIAL  avgt   30  0.049 ± 0.002   s/op
 
BenchmarkBatchStreamReaders.readBlocks             real         NONE  avgt   30  0.048 ± 0.001   s/op
 
BenchmarkBatchStreamReaders.readBlocks           double      PARTIAL  avgt   30  0.062 ± 0.002   s/op
 
BenchmarkBatchStreamReaders.readBlocks           double         NONE  avgt   30  0.070 ± 0.002   s/op
 
 
 
 
 
---- without ----
 
# Run complete. Total time: 00:03:34
 
 
 
Benchmark                             (typeSignature)  (withNulls)  Mode  Cnt  Score   Error  Units
 
BenchmarkSelectiveStreamReaders.read             real         true  avgt   20  0.123 ± 0.003   s/op
 
BenchmarkSelectiveStreamReaders.read             real        false  avgt   20  0.202 ± 0.005   s/op
 
BenchmarkSelectiveStreamReaders.read           double         true  avgt   20  0.126 ± 0.004   s/op
 
BenchmarkSelectiveStreamReaders.read           double        false  avgt   20  0.218 ± 0.013   s/op
 
 
 
# Run complete. Total time: 00:03:34
 
 
 
Benchmark                               (typeSignature)  (withNulls)  Mode  Cnt  Score   Error  Units
 
BenchmarkBatchStreamReaders.readBlocks             real      PARTIAL  avgt   20  0.127 ± 0.003   s/op
 
BenchmarkBatchStreamReaders.readBlocks             real         NONE  avgt   20  0.200 ± 0.007   s/op
 
BenchmarkBatchStreamReaders.readBlocks           double      PARTIAL  avgt   20  0.133 ± 0.006   s/op
 
BenchmarkBatchStreamReaders.readBlocks           double         NONE  avgt   20  0.203 ± 0.006   s/op
 

Next time, let's break up such changes into multiple commits. Here, I'd do the following:

*	commit with OrcInputStream optimizations to remove current Slice.
*	commit with the new readDouble method and corresponding changes in DoubleInputStream
*	commit with the new readFloat method and corresponding changes in FloadInputStream
*	all the changes to optimize reading longs

This change affects default scan path, hence, we need to (1) add release nodes to PR description; (2) run verifier for different types of production workload to make sure nothing breaks.

  _____  

In presto-orc/src/main/java/com/facebook/presto/orc/stream/OrcInputStream.java <https://github.com/prestodb/presto/pull/13434#discussion_r327177962> :

> @@ -71,12 +89,16 @@ public OrcInputStream(
 
         systemMemoryContext.newLocalMemoryContext(OrcInputStream.class.getSimpleName()).setBytes(sliceInputRetainedSizeInBytes);
 
 
 
         if (!decompressor.isPresent()) {
 
-            this.current = sliceInput;
 
+            long sliceInputPosition = sliceInput.position();
 

*	sliceInput.position() and sliceInput.remaining() is called multiple times in this block; I suggest to re-write it to use these only once
*	I don't understand why you call sliceInput.setPosition(sliceInputPosition); - Would you explain?

            this.length = toIntExact(sliceInput.remaining());
 
            this.buffer = new byte[this.length];
 
 
 
            int sliceInputPosition = toIntExact(sliceInput.position());
 
            sliceInput.readFully(buffer, sliceInputPosition, this.length);
 
            sliceInput.setPosition(sliceInputPosition);
 
            
 
            this.compressedSliceInput = EMPTY_SLICE.getInput();
 
  _____  


In presto-orc/src/main/java/com/facebook/presto/orc/stream/OrcInputStream.java <https://github.com/prestodb/presto/pull/13434#discussion_r327184219> :

> +    public void skipVarints(long items)
 
+            throws IOException
 
+    {
 
+        if (items == 0) {
 
+            return;
 
+        }
 
+
 
+        while (items > 0) {
 
+            items -= skipVarintsInBuffer(items);
 
+        }
 
+    }
 
+
 
+    private long skipVarintsInBuffer(long items)
 
+            throws IOException
 
+    {
 
+        if (items == 0) {
 

this check is always false; let's remove it

  _____  

In presto-orc/src/main/java/com/facebook/presto/orc/stream/OrcInputStream.java <https://github.com/prestodb/presto/pull/13434#discussion_r327185132> :

> +    }
 
+
 
+    public float readFloat()
 
+            throws IOException
 
+    {
 
+        int readPosition = ensureContiguousBytesAndAdvance(SIZE_OF_FLOAT);
 
+        if (readPosition < 0) {
 
+            return ByteArrays.getFloat(tempDouble, 0);
 
+        }
 
+        return ByteArrays.getFloat(buffer, readPosition);
 
+    }
 
+
 
+    private int ensureContiguousBytesAndAdvance(int numBytes)
 
+            throws IOException
 
+    {
 
+        // If there are numBytes in the buffer, return the offsett of the start and advances by numBytes. If not, copies numBytes into tempDouble and advamces numBytes and returns -1.
 

typos: offsett -> offset, advances -> advance; copies -> copy; advamces -> advance by; etc.

// If there are numBytes in the buffer, return the offset of the start and advance by numBytes. If not, copy numBytes into tempDouble, advance by numBytes and return -1.
 

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub <https://github.com/prestodb/presto/pull/13434?email_source=notifications&email_token=AKPPPTYIIQUVS5ET6XILWVTQLDSZHA5CNFSM4IZGIQ72YY3PNVWWK3TUL52HS4DFWFIHK3DMKJSXC5LFON2FEZLWNFSXPKTDN5WW2ZLOORPWSZGOCFSW6UA#pullrequestreview-291860304> , or mute the thread <https://github.com/notifications/unsubscribe-auth/AKPPPT5SVV2SVEWVXNSTCOTQLDSZHANCNFSM4IZGIQ7Q> .